Refactor: Split flight search function
Split the flight search function into smaller files for better readability and maintainability.
